requirements.txt
mpxj>=12.0.0
psycopg[binary]>=3.2.0
boto3>=1.34.0
python-dotenv>=1.0.1
